chore: fix darwin build failure
