+++ pkg/azuredisk/azure_common_linux_test.go
