--- pkg/azuredisk/azure_common_unix.go
@@ -1,98 +1,0 @@
-//go:build darwin || linux
-// +build darwin linux
-
-/*
-Copyright 2017 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package azuredisk
-
-import (
-	"fmt"
-	"path/filepath"
-	"strconv"
-	"strings"
-
-	"k8s.io/klog/v2"
-	mount "k8s.io/mount-utils"
-	"sigs.k8s.io/azuredisk-csi-driver/pkg/azureutils"
-)
-
-const sysClassBlockPath = "/sys/class/block/"
-
-func getDevicePathWithMountPath(mountPath string, m *mount.SafeFormatAndMount) (string, error) {
-	args := []string{"-o", "source", "--noheadings", "--mountpoint", mountPath}
-	output, err := m.Exec.Command("findmnt", args...).Output()
-	if err != nil {
-		return "", fmt.Errorf("could not determine device path(%s), error: %v", mountPath, err)
-	}
-
-	devicePath := strings.TrimSpace(string(output))
-	if len(devicePath) == 0 {
-		return "", fmt.Errorf("could not get valid device for mount path: %q", mountPath)
-	}
-
-	return devicePath, nil
-}
-
-func getBlockSizeBytes(devicePath string, m *mount.SafeFormatAndMount) (int64, error) {
-	output, err := m.Exec.Command("blockdev", "--getsize64", devicePath).Output()
-	if err != nil {
-		return -1, fmt.Errorf("error when getting size of block volume at path %s: output: %s, err: %v", devicePath, string(output), err)
-	}
-	strOut := strings.TrimSpace(string(output))
-	gotSizeBytes, err := strconv.ParseInt(strOut, 10, 64)
-	if err != nil {
-		return -1, fmt.Errorf("failed to parse size %s into int a size", strOut)
-	}
-	return gotSizeBytes, nil
-}
-
-func resizeVolume(devicePath, volumePath string, m *mount.SafeFormatAndMount) error {
-	_, err := mount.NewResizeFs(m.Exec).Resize(devicePath, volumePath)
-	return err
-}
-
-// needResizeVolume check whether device needs resize
-func needResizeVolume(devicePath, volumePath string, m *mount.SafeFormatAndMount) (bool, error) {
-	return mount.NewResizeFs(m.Exec).NeedResize(devicePath, volumePath)
-}
-
-// rescanVolume rescan device for detecting device size expansion
-// devicePath e.g. `/dev/sdc`
-func rescanVolume(io azureutils.IOHandler, devicePath string) error {
-	klog.V(6).Infof("rescanVolume - begin to rescan %s", devicePath)
-	deviceName := filepath.Base(devicePath)
-	rescanPath := filepath.Join(sysClassBlockPath, deviceName, "device/rescan")
-	return io.WriteFile(rescanPath, []byte("1"), 0666)
-}
-
-// rescanAllVolumes rescan all sd* devices under /sys/class/block/sd* starting from sdc
-func rescanAllVolumes(io azureutils.IOHandler) error {
-	dirs, err := io.ReadDir(sysClassBlockPath)
-	if err != nil {
-		return err
-	}
-	for _, device := range dirs {
-		deviceName := device.Name()
-		if strings.HasPrefix(deviceName, "sd") && deviceName >= "sdc" {
-			path := filepath.Join(sysClassBlockPath, deviceName)
-			if err := rescanVolume(io, path); err != nil {
-				klog.Warningf("rescanVolume - rescan %s failed with %v", path, err)
-			}
-		}
-	}
-	return nil
-}
